refactoring to avoid duplicate log lines
